+++ source/ProcessManager.Example.Orchestrations.Tests/ProcessManager.Example.Orchestrations.Tests.csproj
@@ -8,6 +8,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+        <Compile Include="..\Shared\Tests\Fixtures\DomainTestDataFactory.cs" Link="Fixtures\DomainTestDataFactory.cs" />
         <Compile Include="..\Shared\Tests\Fixtures\Extensions\ServiceBusReceivedMessageExtensions.cs" Link="Fixtures\Extensions\ServiceBusReceivedMessageExtensions.cs" />
         <Compile Include="..\Shared\Tests\Fixtures\AuthenticationOptionsForTests.cs" Link="Fixtures\AuthenticationOptionsForTests.cs" />
         <Compile Include="..\Shared\Tests\Fixtures\ExampleOrchestrationsAppManager.cs" Link="Fixtures\ExampleOrchestrationsAppManager.cs" />
