+++ source/ProcessManager.Orchestrations.Tests/Integration/CustomQueries/Calculations/V1/SearchCalculationsHandlerV1Tests.cs
@@ -22,10 +22,10 @@ using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.CustomQueries
 using Energinet.DataHub.ProcessManager.Orchestrations.CustomQueries.Calculations.V1;
 using Energinet.DataHub.ProcessManager.Orchestrations.Tests.Fixtures;
 using Energinet.DataHub.ProcessManager.Shared.Api.Mappers;
+using Energinet.DataHub.ProcessManager.Shared.Tests.Fixtures;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.SqlServer.NodaTime.Extensions;
-using Moq;
 using NodaTime;
 using ApiModel = Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 
@@ -1022,7 +1022,7 @@ public class SearchCalculationsHandlerV1Tests :
             Instant isTerminatedAsSucceededAt = default)
     {
         var orchestrationDescription = builder.Build();
-        var orchestrationInstances = CreateLifecycleDataset(
+        var orchestrationInstances = DomainTestDataFactory.CreateLifecycleDataset(
             orchestrationDescription,
             isRunningStartedAt,
             isTerminatedAsSucceededAt);
@@ -1037,94 +1037,6 @@ public class SearchCalculationsHandlerV1Tests :
         await dbContext.SaveChangesAsync();
 
         return orchestrationInstances;
-    }
-
-    /// <summary>
-    /// Create orchestration instances from <paramref name="orchestrationDescription"/>
-    /// in the following lifecycle states:
-    ///  - Pending
-    ///  - Queued
-    ///  - Running
-    ///  - Terminated as succeeded
-    ///  - Terminated as failed
-    ///
-    /// If <paramref name="isRunningStartedAt"/> is specified, then this value
-    /// is used when transitioning to Running.
-    ///
-    /// If <paramref name="isTerminatedAsSucceededAt"/> is specified, then this value
-    /// is used when transitioning to terminated as succeeded.
-    /// </summary>
-    private (
-            OrchestrationInstance IsPending,
-            OrchestrationInstance IsQueued,
-            OrchestrationInstance IsRunning,
-            OrchestrationInstance IsTerminatedAsSucceeded,
-            OrchestrationInstance IsTerminatedAsFailed)
-        CreateLifecycleDataset(
-            OrchestrationDescription orchestrationDescription,
-            Instant isRunningStartedAt = default,
-            Instant isTerminatedAsSucceededAt = default)
-    {
-        var isPending = OrchestrationInstance.CreateFromDescription(
-            identity: _userIdentity.MapToDomain(),
-            description: orchestrationDescription,
-            skipStepsBySequence: [],
-            clock: SystemClock.Instance);
-
-        var isQueued = OrchestrationInstance.CreateFromDescription(
-            identity: _userIdentity.MapToDomain(),
-            description: orchestrationDescription,
-            skipStepsBySequence: [],
-            clock: SystemClock.Instance);
-        isQueued.Lifecycle.TransitionToQueued(SystemClock.Instance);
-
-        var isRunning = OrchestrationInstance.CreateFromDescription(
-            identity: _userIdentity.MapToDomain(),
-            description: orchestrationDescription,
-            skipStepsBySequence: [],
-            clock: SystemClock.Instance);
-        isRunning.Lifecycle.TransitionToQueued(SystemClock.Instance);
-        if (isRunningStartedAt == default)
-        {
-            isRunning.Lifecycle.TransitionToRunning(SystemClock.Instance);
-        }
-        else
-        {
-            var clockMock = new Mock<IClock>();
-            clockMock.Setup(m => m.GetCurrentInstant())
-                .Returns(isRunningStartedAt);
-            isRunning.Lifecycle.TransitionToRunning(clockMock.Object);
-        }
-
-        var isTerminatedAsSucceeded = OrchestrationInstance.CreateFromDescription(
-            identity: _userIdentity.MapToDomain(),
-            description: orchestrationDescription,
-            skipStepsBySequence: [],
-            clock: SystemClock.Instance);
-        isTerminatedAsSucceeded.Lifecycle.TransitionToQueued(SystemClock.Instance);
-        isTerminatedAsSucceeded.Lifecycle.TransitionToRunning(SystemClock.Instance);
-        if (isTerminatedAsSucceededAt == default)
-        {
-            isTerminatedAsSucceeded.Lifecycle.TransitionToSucceeded(SystemClock.Instance);
-        }
-        else
-        {
-            var clockMock = new Mock<IClock>();
-            clockMock.Setup(m => m.GetCurrentInstant())
-                .Returns(isTerminatedAsSucceededAt);
-            isTerminatedAsSucceeded.Lifecycle.TransitionToSucceeded(clockMock.Object);
-        }
-
-        var isTerminatedAsFailed = OrchestrationInstance.CreateFromDescription(
-            identity: _userIdentity.MapToDomain(),
-            description: orchestrationDescription,
-            skipStepsBySequence: [],
-            clock: SystemClock.Instance);
-        isTerminatedAsFailed.Lifecycle.TransitionToQueued(SystemClock.Instance);
-        isTerminatedAsFailed.Lifecycle.TransitionToRunning(SystemClock.Instance);
-        isTerminatedAsFailed.Lifecycle.TransitionToFailed(SystemClock.Instance);
-
-        return (isPending, isQueued, isRunning, isTerminatedAsSucceeded, isTerminatedAsFailed);
     }
 
     /// <summary>
@@ -1145,7 +1057,7 @@ public class SearchCalculationsHandlerV1Tests :
             canBeScheduled: true,
             functionName: "TestOrchestrationFunction");
 
-        var johnDoe = CreateLifecycleDataset(johnDoeV1Description);
+        var johnDoe = DomainTestDataFactory.CreateLifecycleDataset(johnDoeV1Description);
 
         var isTerminatedAsUserCancelled = OrchestrationInstance.CreateFromDescription(
             identity: _userIdentity.MapToDomain(),
